# server/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OmniSave",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/sync/transactions/{transaction_id}/range": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Download Range",
        "operationId": "download_range_api_v1_sync_transactions__transaction_id__range_get",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offset"
            }
          },
          {
            "name": "length",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Length"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/ack": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Ack Restore",
        "operationId": "ack_restore_api_v1_sync_ack_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AckBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/fail": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Delivery Fail",
        "description": "Permanent inject failure \u2014 marks outbound FAILED, removed from queue.",
        "operationId": "delivery_fail_api_v1_sync_fail_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FailBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AckBody": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          }
        },
        "type": "object",
        "required": [
          "transaction_id"
        ],
        "title": "AckBody"
      },
...
      "FailBody": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "error_code": {
            "type": "string",
            "title": "Error Code",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "transaction_id"
        ],
        "title": "FailBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}